Fixing send/poll race
Fixing a bug where the server would throw if send request is received before the first poll or sse request (#57). We will now delay send until the first poll or sse is received or reject the send request if poll or sse not received within a timeout
Also making the connection not to be marked as Active upon reserving a connection. This enables cleaning up connections that were reserved but never used.