Change the size header to little endian
While it has been traditional to use Big Endian byte orders in protocols, it really is only that a tradition. As almost all platforms (ARM can run in both in theory, but isn't normally configured for anything other than Little, and I think the XBOX) will run in little endian is there a good reason to burn ops/the hassle of bit flipping?