change to use Karma for Functional Tests
This removes a bunch of our custom logic and makes it easier to use
SauceLabs. The core difference is that we server the Functional Test
JavaScript from a **different** server than the SignalR Hub. This means
the test are always running cross-origin, but that's OK. This only
happens in the command-line and SauceLabs build. You can still run the
Functional Tests in the browser on the same origin.